Add test_cli_profile; Various auto-pepping
