+++ tests/test_profile.py
@@ -134,17 +134,6 @@ def test_BasicDatasetProfiler_with_context(empty_data_context, filesystem_csv_2)
     }
 
 
-@pytest.fixture()
-def filesystem_csv_2(tmp_path_factory):
-    base_dir = tmp_path_factory.mktemp('test_files')
-    base_dir = str(base_dir)
-
-    # Put a file in the directory
-    toy_dataset = PandasDataset({"x": [1, 2, 3]})
-    toy_dataset.to_csv(os.path.join(base_dir, "f1.csv"), index=None)
-    return base_dir
-
-
 def test_context_profiler(empty_data_context, filesystem_csv_2):
     empty_data_context.add_datasource(
         "my_datasource", "pandas", base_directory=str(filesystem_csv_2))
